add activity log

+++ application/modules/admin/controllers/Products.php
@@ -9,53 +9,73 @@ class Products extends MX_Controller
         parent::__construct();
         $this->load->model("m_products");
         $this->load->library('form_validation');
+        $this->load->model('login/m_login');
     }
 
     public function index()
     {
-        $data["products"] = $this->m_products->getAll();
-        $this->load->view("admin/products/list", $data);
+        if($this->m_login->cek_session())
+        {
+            $data["products"] = $this->m_products->getAll();
+            $this->load->view("admin/products/list", $data);
+        }else{
+            //jika session belum terdaftar, maka redirect ke halaman login
+            redirect(base_url('login'));
+        }
     }
 
     public function add()
     {
-        $product = $this->m_products;
-        $validation = $this->form_validation;
-        $validation->set_rules($product->rules());
+        if($this->m_login->cek_session()){
+            $product = $this->m_products;
+            $validation = $this->form_validation;
+            $validation->set_rules($product->rules());
 
-        if ($validation->run()) {
-            $product->save();
-            $this->session->set_flashdata('success', 'Berhasil disimpan');
+            if ($validation->run()) {
+                $product->save();
+                $this->session->set_flashdata('success', 'Berhasil disimpan');
+            }
+            
+            $this->load->view("admin/products/new_form");
+        }else{
+            //jika session belum terdaftar, maka redirect ke halaman login
+            redirect(base_url('login'));
         }
-
-        $this->load->view("admin/products/new_form");
     }
 
-    public function edit($id = null)
-    {
-        if (!isset($id)) redirect('admin/products');
-       
-        $product = $this->m_products;
-        $validation = $this->form_validation;
-        $validation->set_rules($product->rules());
-
-        if ($validation->run()) {
-            $product->update();
-            $this->session->set_flashdata('success', 'Berhasil disimpan');
+    public function edit($id = null){
+        if($this->m_login->cek_session()){
+            if (!isset($id)) redirect('admin/products');
+           
+            $product = $this->m_products;
+            $validation = $this->form_validation;
+            $validation->set_rules($product->rules());
+    
+            if ($validation->run()) {
+                $product->update();
+                $this->session->set_flashdata('success', 'Berhasil disimpan');
+            }
+    
+            $data["product"] = $product->getById($id);
+            if (!$data["product"]) show_404();
+            
+            $this->load->view("admin/products/edit_form", $data);
+        }else{
+            //jika session belum terdaftar, maka redirect ke halaman login
+            redirect(base_url('login'));
         }
-
-        $data["product"] = $product->getById($id);
-        if (!$data["product"]) show_404();
-        
-        $this->load->view("admin/products/edit_form", $data);
     }
 
-    public function delete($id=null)
-    {
-        if (!isset($id)) show_404();
-        
-        if ($this->m_products->delete($id)) {
-            redirect(base_url('admin/products'));
+    public function delete($id=null){
+        if($this->m_login->cek_session()){
+            if (!isset($id)) show_404();
+            
+            if ($this->m_products->delete($id)) {
+                redirect(base_url('admin/products'));
+            }
+        }else{
+            //jika session belum terdaftar, maka redirect ke halaman login
+            redirect(base_url('login'));
         }
     }
 }
